requirements.txt
numpy
matplotlib
nilearn
nibabel
scikit-learn
torch
torchvision
pillow